% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_config.R
\name{as_config}
\alias{as_config}
\title{Coerce a config list to a config class object}
\usage{
as_config(x)
}
\arguments{
\item{x}{a list representation of the contents a \code{tasks.json} config file.}
}
\value{
a config list object with subclass \verb{<config>}.
}
\description{
Coerce a config list to a config class object
}
\examples{
config_tasks <- read_config(
  hub_path = system.file("testhubs/simple", package = "hubUtils")
)
# Remove all attributes except names to demonstrate functionality
attributes(config_tasks) <- attributes(config_tasks)[
  names(attributes(config_tasks)) == "names"
]
# Convert to config object
as_config(config_tasks)
}
