% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_arrow_schema.R
\name{build_arrow_schema}
\alias{build_arrow_schema}
\title{Create an arrow schema}
\usage{
build_arrow_schema(
  config_tasks,
  format = c("csv", "parquet", "error"),
  partitions = list(team = arrow::utf8())
)
}
\arguments{
\item{config_tasks}{a list version of the content's of a hub's \code{tasks.json}
config file created using function \code{\link[=read_config]{read_config()}}.}

\item{format}{the hub file format.}

\item{partitions}{a named list specifying the arrow data types of any
partitioning column. Only relevant for opening \code{"parquet"} format datasets.
Ignored for all other formats.}
}
\value{
an arrow schema object that can be used to define column datatypes when
opening model output data
}
\description{
Create an arrow schema from a the \code{tasks.json} config file for using when
opening an arrow dataset.
}
\examples{
origin_path <- system.file("testhubs/simple", package = "hubUtils")
config_tasks <- hubUtils:::read_config(origin_path, "tasks")
schema_csv <- build_arrow_schema(config_tasks, format = "csv")
con_csv <- arrow::open_dataset(
    origin_path, format = "csv",
    partitioning = "team",
    col_types = schema_csv,
    factory_options = list(exclude_invalid_files = TRUE))
con_csv

schema_parquet <- build_arrow_schema(config_tasks, format = "parquet")
con_parquet <- arrow::open_dataset(
    origin_path, format = "parquet",
    partitioning = "team",
    schema = schema_parquet,
    factory_options = list(exclude_invalid_files = TRUE))
con_parquet

hub_con <- arrow::open_dataset(
    sources = list(
        con_csv,
        con_parquet)
)
hub_con
}
