% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub-connection.R
\name{connect_hub}
\alias{connect_hub}
\title{Connect to a Modeling Hub}
\usage{
connect_hub(
  hub_path,
  hubmeta_path = NULL,
  hubmeta_format = c("json", "yaml", "yml")
)
}
\arguments{
\item{hub_path}{Path to the Modeling Hub directory.}

\item{hubmeta_path}{Path to a hubmeta file. The default (\code{NULL}) expects a file
in the root of the directory given in \code{hub_path} named \code{hubmeta} and having
an accepted file format (see \code{hubmeta_format}). Use this argument if the
hubmeta file is in a non-default location or has a non-default file name and
include the file extension in the path.}

\item{hubmeta_format}{The file extension of the hubmeta file. One of \code{"json"},
\code{"yaml"} or \code{"yml"}. Defaults to \code{"json"} and ignored when \code{hubmeta_path} is
provided.}
}
\value{
an S3 object of class \verb{<hub-connection>} containing metadata about the
modeling hub. This object is used to communicate with the Modelling Hub when
extracting or submitting data.
}
\description{
Connect to a Modeling Hub
}
\examples{
# Simple forecasting Hub example
hub_path <- system.file("hub_1", package = "hubUtils")
connect_hub(hub_path)
# Scenario Hub example
scnr_path <- system.file("scnr_hub_1", package = "hubUtils")
connect_hub(hub_path)
}
