% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hub_schema.R
\name{create_hub_schema}
\alias{create_hub_schema}
\title{Create a Hub arrow schema}
\usage{
create_hub_schema(
  config_tasks,
  format = c("csv", "parquet", "arrow"),
  partitions = list(model_id = arrow::utf8()),
  output_type_id_datatype = c("auto", "character", "double", "integer", "logical",
    "Date")
)
}
\arguments{
\item{config_tasks}{a list version of the content's of a hub's \code{tasks.json}
config file created using function \code{\link[=read_config]{read_config()}}.}

\item{format}{the hub file format.}

\item{partitions}{a named list specifying the arrow data types of any
partitioning column. Only relevant for opening \code{"parquet"} or \code{"arrow"} format datasets.
Ignored for all other formats.}

\item{output_type_id_datatype}{character string. One of \code{"auto"}, \code{"character"},
\code{"double"}, \code{"integer"}, \code{"logical"}, \code{"Date"}. Defaults to \code{"auto"} indicating
that \code{output_type_id} will be determined automatically from the \code{tasks.json}
config file. Other data type values can be used to override automatic determination.
Note that attempting to coerce \code{output_type_id} to a data type that is not possible
(e.g. trying to coerce to \code{"double"} when the data contains \code{"character"} values)
will likely result in an error or potentially unexpected behaviour so use with
care.}
}
\value{
an arrow schema object that can be used to define column datatypes when
opening model output data. If the schema is being created for a \code{parquet}/\code{arrow} hub,
the partitioning variable data type is included in the schema while for \code{csv} hubs
the partitioning variable is not as it causes errors when opening.
}
\description{
Create an arrow schema from a \code{tasks.json} config file. For use when
opening an arrow dataset.
}
\examples{
hub_path <- system.file("testhubs/simple", package = "hubUtils")
config_tasks <- read_config(hub_path, "tasks")
schema_csv <- create_hub_schema(config_tasks, format = "csv")
schema_parquet <- create_hub_schema(config_tasks, format = "parquet")
}
