% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_type.R
\name{create_output_type}
\alias{create_output_type}
\title{Create an \code{output_type} class object.}
\usage{
create_output_type(...)
}
\arguments{
\item{...}{objects of class \code{output_type_item} created using functions from the
\verb{create_output_type_*()} family of functions.}
}
\value{
a named list of class \code{output_type}.
}
\description{
Create an \code{output_type} class object.
}
\details{
For more details consult
the \href{https://hubdocs.readthedocs.io/en/latest/format/hub-metadata.html#hub-model-task-metadata-tasks-json-file}{documentation on \code{tasks.json} Hub config files}.
}
\examples{
create_output_type(
  create_output_type_mean(
    is_required = TRUE,
    value_type = "double",
    value_minimum = 0L
  ),
  create_output_type_median(
    is_required = FALSE,
    value_type = "double"
  ),
  create_output_type_quantile(
    required = c(0.25, 0.5, 0.75),
    optional = c(
      0.1, 0.2, 0.3, 0.4, 0.6,
      0.7, 0.8, 0.9
    ),
    value_type = "double",
    value_minimum = 0
  )
)
}
\seealso{
\code{\link[=create_output_type_mean]{create_output_type_mean()}}, \code{\link[=create_output_type_median]{create_output_type_median()}},
\code{\link[=create_output_type_quantile]{create_output_type_quantile()}}, \code{\link[=create_output_type_cdf]{create_output_type_cdf()}},
\code{\link[=create_output_type_pmf]{create_output_type_pmf()}}, \code{\link[=create_output_type_sample]{create_output_type_sample()}}
}
