% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_type_item.R
\name{create_output_type_quantile}
\alias{create_output_type_quantile}
\alias{create_output_type_cdf}
\alias{create_output_type_pmf}
\alias{create_output_type_sample}
\title{Create a distribution output type object of class \code{output_type_item}}
\usage{
create_output_type_quantile(
  required,
  optional,
  value_type,
  value_minimum = NULL,
  value_maximum = NULL,
  schema_version = "latest",
  branch = "main"
)

create_output_type_cdf(
  required,
  optional,
  value_type,
  schema_version = "latest",
  branch = "main"
)

create_output_type_pmf(
  required,
  optional,
  value_type,
  schema_version = "latest",
  branch = "main"
)

create_output_type_sample(
  required,
  optional,
  value_type,
  value_minimum = NULL,
  value_maximum = NULL,
  schema_version = "latest",
  branch = "main"
)
}
\arguments{
\item{required}{Atomic vector of required \code{output_type_id} values. Can be NULL if
all values are optional.}

\item{optional}{Atomic vector of optional \code{output_type_id} values. Can be NULL if
all values are required.}

\item{value_type}{Character string. The data type of the output_type values.}

\item{value_minimum}{Numeric. The inclusive minimum of output_type values.}

\item{value_maximum}{Numeric. The inclusive maximum of output_type values.}

\item{schema_version}{Character string specifying the json schema version to
be used for validation. The default value \code{"latest"} will use the latest version
available in the Infectious Disease Modeling Hubs
\href{https://github.com/Infectious-Disease-Modeling-Hubs/schemas}{schemas repository}.
Alternatively, a specific version of a schema (e.g. \code{"v0.0.1"})  can be
specified.}

\item{branch}{The branch of the Infectious Disease Modeling Hubs
\href{https://github.com/Infectious-Disease-Modeling-Hubs/schemas}{schemas repository}
from which to fetch schema. Defaults to \code{"main"}.}
}
\value{
a named list of class \code{output_type_item} representing a \code{quantile},
\code{cdf}, \code{pmf} or \code{sample} output type.
}
\description{
Create a representation of a \code{quantile}, \code{cdf}, \code{pmf} or \code{sample} output
type as a list object of class \code{output_type_item}. This can be combined with
additional \code{output_type_item}s using function \code{\link[=create_output_type]{create_output_type()}} to
create an \code{output_type} object for a given model_task.
This can be combined with other building blocks which can then be written as
or appended to \code{tasks.json} Hub config files.
output type.
This can be combined with other building blocks which can then be written as
or appended to \code{tasks.json} Hub config files.
}
\details{
For more details consult
the \href{https://hubdocs.readthedocs.io/en/latest/format/hub-metadata.html#hub-model-task-metadata-tasks-json-file}{documentation on \code{tasks.json} Hub config files}.
}
\section{Functions}{
\itemize{
\item \code{create_output_type_quantile()}: Create a list representation of a \code{quantile}
output type.

\item \code{create_output_type_cdf()}: Create a list representation of a \code{cdf}
output type.

\item \code{create_output_type_pmf()}: Create a list representation of a \code{pmf}
output type.

\item \code{create_output_type_sample()}: Create a list representation of a \code{sample}
output type.

}}
\examples{
create_output_type_quantile(
  required = c(0.25, 0.5, 0.75),
  optional = c(
    0.1, 0.2, 0.3, 0.4, 0.6,
    0.7, 0.8, 0.9
  ),
  value_type = "double",
  value_minimum = 0
)
create_output_type_cdf(
  required = c(10, 20),
  optional = NULL,
  value_type = "double"
)
create_output_type_cdf(
  required = NULL,
  optional = c("EW202240", "EW202241", "EW202242"),
  value_type = "double"
)
create_output_type_pmf(
  required = NULL,
  optional = c("low", "moderate", "high", "extreme"),
  value_type = "double"
)
create_output_type_sample(
  required = 1:10, optional = 11:15,
  value_type = "double"
)
}
\seealso{
\code{\link[=create_output_type]{create_output_type()}}
}
