% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_round.R
\name{create_round}
\alias{create_round}
\title{Create an object of class \code{round}}
\usage{
create_round(
  round_id_from_variable,
  round_id,
  round_name = NULL,
  model_tasks,
  submissions_due,
  last_data_date = NULL,
  file_format = NULL
)
}
\arguments{
\item{round_id_from_variable}{logical. Whether \code{round_id} is inferred from the
values of a \code{task_id} variable within the \code{model_tasks} \code{model_task} items.}

\item{round_id}{character string. The round identifier.
If \code{round_id_from_variable = TRUE}, \code{round_id} should be the name of a \code{task_id}
variable present in all  \code{model_tasks} \code{model_task} items.}

\item{round_name}{character string. An optional round name. This can be useful
for internal referencing of rounds, for examples, when a date is used as
\code{round_id} but hub maintainers and teams also refer to rounds as round-1, round-2 etc.}

\item{model_tasks}{an object of class \code{model_tasks} created with function
\code{\link[=create_model_tasks]{create_model_tasks()}}.}

\item{submissions_due}{named list conforming to one of the two following structures:
\enumerate{
\item Submission dates for round is determined relative to an origin date.
\itemize{
\item \code{relative_to}: character string of the name of the \code{task_id} variable containing
origin dates in relation to which submission start and end dates are determined.
\item \code{start}: integer. Difference in days between start and origin date.
\item \code{end}: integer. Difference in days between end and origin date.
}
\item Submission dates for round are provided explicitly.
\itemize{
\item \code{start}: character. Submission start date in ISO 8601 format (YYYY-MM-DD).
\item \code{end}: character. Submission end date in ISO 8601 format (YYYY-MM-DD).
}
}}

\item{last_data_date}{character date in ISO 8601 format (YYYY-MM-DD).
The last date with recorded data in the data set used as input to a model.
Optional.}

\item{file_format}{character string. An optional specification of a file format for the round.
This option in only available for some versions of the schema and is ignored if
not allowed in the version of the schema used. It also overrides any specification
of file format in \code{admin.json}. For more details on whether this argument can
be used as well as available formats, please consult
the \href{https://hubdocs.readthedocs.io/en/latest/format/hub-metadata.html#hub-model-task-metadata-tasks-json-file}{documentation on \code{tasks.json} Hub config files}.}
}
\value{
a named list of class \code{round}.
}
\description{
Create a representation of a round item as a list object of
class \code{round}. This can be combined with
additional \code{round} objects using function \code{\link[=create_rounds]{create_rounds()}}.
Such building blocks can ultimately be combined and then written out as or
appended to \code{tasks.json} Hub config files.
}
\details{
For more details consult
the \href{https://hubdocs.readthedocs.io/en/latest/format/hub-metadata.html#hub-model-task-metadata-tasks-json-file}{documentation on \code{tasks.json} Hub config files}.
}
\examples{
model_tasks <- create_model_tasks(
  create_model_task(
    task_ids = create_task_ids(
      create_task_id("origin_date",
        required = NULL,
        optional = c(
          "2023-01-02",
          "2023-01-09",
          "2023-01-16"
        )
      ),
      create_task_id("location",
        required = "US",
        optional = c("01", "02", "04", "05", "06")
      ),
      create_task_id("horizon",
        required = 1L,
        optional = 2:4
      )
    ),
    output_type = create_output_type(
      create_output_type_mean(
        is_required = TRUE,
        value_type = "double",
        value_minimum = 0L
      )
    ),
    target_metadata = create_target_metadata(
      create_target_metadata_item(
        target_id = "inc hosp",
        target_name = "Weekly incident influenza hospitalizations",
        target_units = "rate per 100,000 population",
        target_keys = NULL,
        target_type = "discrete",
        is_step_ahead = TRUE,
        time_unit = "week"
      )
    )
  )
)
create_round(
  round_id_from_variable = TRUE,
  round_id = "origin_date",
  model_tasks = model_tasks,
  submissions_due = list(
    relative_to = "origin_date",
    start = -4L,
    end = 2L
  ),
  last_data_date = "2023-01-02"
)
}
\seealso{
\code{\link[=create_rounds]{create_rounds()}}
}
