% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_target_metadata.R
\name{create_target_metadata}
\alias{create_target_metadata}
\title{Create a \code{target_metadata} class object.}
\usage{
create_target_metadata(...)
}
\arguments{
\item{...}{objects of class \code{target_metadata_item} created using function
\code{\link[=create_target_metadata_item]{create_target_metadata_item()}}}
}
\value{
a named list of class \code{target_metadata}.
}
\description{
Create a \code{target_metadata} class object.
}
\details{
For more details consult
the \href{https://hubdocs.readthedocs.io/en/latest/format/hub-metadata.html#hub-model-task-metadata-tasks-json-file}{documentation on \code{tasks.json} Hub config files}.
}
\examples{
create_target_metadata(
  create_target_metadata_item(
    target_id = "inc hosp",
    target_name = "Weekly incident influenza hospitalizations",
    target_units = "rate per 100,000 population",
    target_keys = list(target = "inc hosp"),
    target_type = "discrete",
    is_step_ahead = TRUE,
    time_unit = "week"
  ),
  create_target_metadata_item(
    target_id = "inc death",
    target_name = "Weekly incident influenza deaths",
    target_units = "rate per 100,000 population",
    target_keys = list(target = "inc death"),
    target_type = "discrete",
    is_step_ahead = TRUE,
    time_unit = "week"
  )
)
}
\seealso{
\code{\link[=create_target_metadata_item]{create_target_metadata_item()}}
}
