% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_target_metadata_item.R
\name{create_target_metadata_item}
\alias{create_target_metadata_item}
\title{Create an object of class \code{target_metadata_item}}
\usage{
create_target_metadata_item(
  target_id,
  target_name,
  target_units,
  target_keys = NULL,
  description = NULL,
  target_type,
  is_step_ahead,
  time_unit = NULL,
  schema_version = "latest",
  branch = "main"
)
}
\arguments{
\item{target_id}{character string. Short description that uniquely identifies
the target.}

\item{target_name}{character string. A longer human readable target description
that could be used, for example, as a visualisation axis label.}

\item{target_units}{character string. Unit of observation of the target.}

\item{target_keys}{named list or \code{NULL}. Should be \code{NULL}, in the case
where the target is not specified as a task_id but is specified solely through
the \code{target_id} argument. Otherwise, should be a named list of one or more
character strings. The name of each element should match a task_id variable
within the same model_tasks object. Each element should be of length 1.
Each value, or the combination of values if multiple keys are specified,
define a single target value.}

\item{description}{character string (optional). An optional verbose description
of the target that might include information such as definitions of a 'rate' or similar.}

\item{target_type}{character string. Target statistical data type. Consult the
\href{https://hubdocs.readthedocs.io/en/latest/format/hub-metadata.html#model-tasks-tasks-json-interactive-schema}{appropriate version of the hub schema}
for potential values.}

\item{is_step_ahead}{logical. Whether the target is part of a sequence of values}

\item{time_unit}{character string. If \code{is_step_ahead} is \code{TRUE}, then this
argument is required and defines the unit of time steps. if \code{is_step_ahead} is
\code{FALSE}, then this argument is not required and will be ignored if given.}

\item{schema_version}{Character string specifying the json schema version to
be used for validation. The default value \code{"latest"} will use the latest version
available in the Infectious Disease Modeling Hubs
\href{https://github.com/Infectious-Disease-Modeling-Hubs/schemas}{schemas repository}.
Alternatively, a specific version of a schema (e.g. \code{"v0.0.1"})  can be
specified.}

\item{branch}{The branch of the Infectious Disease Modeling Hubs
\href{https://github.com/Infectious-Disease-Modeling-Hubs/schemas}{schemas repository}
from which to fetch schema. Defaults to \code{"main"}.}
}
\value{
a named list of class \code{target_metadata_item}.
}
\description{
Create a representation of a target_metadata item as a list object of
class \code{target_metadata_item}. This can be combined with
additional target_metadata items using function \code{\link[=create_target_metadata]{create_target_metadata()}} to
create a target_metadata object for a given model_task.
Such building blocks can ultimately be combined and then written out as or
appended to \code{tasks.json} Hub config files.
}
\details{
For more details consult
the \href{https://hubdocs.readthedocs.io/en/latest/format/hub-metadata.html#hub-model-task-metadata-tasks-json-file}{documentation on \code{tasks.json} Hub config files}.
}
\examples{
create_target_metadata_item(
  target_id = "inc hosp",
  target_name = "Weekly incident influenza hospitalizations",
  target_units = "rate per 100,000 population",
  target_keys = list(target = "inc hosp"),
  target_type = "discrete",
  is_step_ahead = TRUE,
  time_unit = "week"
)
}
\seealso{
\code{\link[=create_target_metadata]{create_target_metadata()}}
}
