% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_task_ids.R
\name{create_task_ids}
\alias{create_task_ids}
\title{Create a \code{task_ids} class object.}
\usage{
create_task_ids(...)
}
\arguments{
\item{...}{objects of class \code{task_id} created using function \code{\link[=create_task_id]{create_task_id()}}}
}
\value{
a named list of class \code{task_ids}.
}
\description{
Create a \code{task_ids} class object.
}
\details{
For more details consult
the \href{https://hubdocs.readthedocs.io/en/latest/format/hub-metadata.html#hub-model-task-metadata-tasks-json-file}{documentation on \code{tasks.json} Hub config files}.
}
\examples{
create_task_ids(
  create_task_id("origin_date",
    required = NULL,
    optional = c(
      "2023-01-02",
      "2023-01-09",
      "2023-01-16"
    )
  ),
  create_task_id("scenario_id",
    required = NULL,
    optional = c(
      "A-2021-03-28",
      "B-2021-03-28"
    )
  ),
  create_task_id("location",
    required = "US",
    optional = c("01", "02", "04", "05", "06")
  ),
  create_task_id("target",
    required = "inc hosp",
    optional = NULL
  ),
  create_task_id("horizon",
    required = 1L,
    optional = 2:4
  )
)
}
\seealso{
\code{\link[=create_task_id]{create_task_id()}}
}
