% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_model_out_val_grid.R
\name{expand_model_out_val_grid}
\alias{expand_model_out_val_grid}
\title{Create expanded grid of valid task ID and output type value combinations}
\usage{
expand_model_out_val_grid(
  config_tasks,
  round_id,
  required_vals_only = FALSE,
  skip_date_coercion = FALSE
)
}
\arguments{
\item{config_tasks}{a list version of the content's of a hub's \code{tasks.json}
config file, accessed through the \code{"config_tasks"} attribute of a \verb{<hub_connection>}
object or function \code{\link[=read_config]{read_config()}}.}

\item{round_id}{Character string. Round identifier. If the round is set to
\code{round_id_from_variable: true}, IDs are values of the task ID defined in the round's
\code{round_id} property of \code{config_tasks}.
Otherwise should match round's \code{round_id} value in config. Ignored if hub
contains only a single round.}

\item{required_vals_only}{Logical. Whether to return only combinations of
Task ID and related output type ID required values.}

\item{skip_date_coercion}{Logical. Whether to skip coercing dates. This can be faster,
especially for larger \code{tbl}s.}
}
\value{
a tibble containing all possible task ID and related output type ID
value combinations. Columns are coerced to data types according to the hub schema,
unless \code{skip_date_coercion = TRUE}. If \code{skip_date_coercion = TRUE}, date columns are returned as
character which can be much faster when large expanded grids are expected.
If \code{required_vals_only = TRUE}, values are limited to the combinations of required
values only.
}
\description{
Create expanded grid of valid task ID and output type value combinations
}
\details{
When a round is set to \code{round_id_from_variable: true},
the value of the task ID from which round IDs are derived (i.e. the task ID
specified in \code{round_id} property of \code{config_tasks}) is set to the value of the
\code{round_id} argument in the returned output.
}
\examples{
hub_con <- connect_hub(
  system.file("testhubs/flusight", package = "hubUtils")
)
config_tasks <- attr(hub_con, "config_tasks")
expand_model_out_val_grid(config_tasks, round_id = "2023-01-02")
expand_model_out_val_grid(
  config_tasks,
  round_id = "2023-01-02",
  required_vals_only = TRUE
)
# Specifying a round in a hub with multiple round configurations.
hub_con <- connect_hub(
  system.file("testhubs/simple", package = "hubUtils")
)
config_tasks <- attr(hub_con, "config_tasks")
expand_model_out_val_grid(config_tasks, round_id = "2022-10-01")
# Later round_id maps to round config that includes additional task ID 'age_group'.
expand_model_out_val_grid(config_tasks, round_id = "2022-10-29")
# Skip coercing dates
expand_model_out_val_grid(config_tasks,
  round_id = "2022-10-29",
  skip_date_coercion = TRUE
)
}
