% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_task_id_cols.R
\name{get_task_id_cols}
\alias{get_task_id_cols}
\title{Get task ID column names from \code{model_out_tbl} object}
\usage{
get_task_id_cols(model_out_tbl)
}
\arguments{
\item{model_out_tbl}{an object of class \code{model_out_tbl}}
}
\value{
a character vector of task ID column names
}
\description{
Get task ID column names from \code{model_out_tbl} object
}
