% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-metadata.R
\name{get_task_id_vals}
\alias{get_task_id_vals}
\title{Get task id values for a given round}
\usage{
get_task_id_vals(con, task_id, round_id = NULL, flatten = TRUE)
}
\arguments{
\item{con}{a \code{hub_connection} class object.}

\item{task_id}{Character string.  A task id name.}

\item{round_id}{Character string.  A round id name. If rounds vary by a variable in hub,
argument is ignored and can be left \code{NULL} (default).}

\item{flatten}{Logical. Whether to flatten required & optional task ids into a single
character vector.}
}
\value{
if \code{flatten = TRUE} (default), a character vector of task id values.
\code{flatten = FALSE}, a named list containing character vectors of
\code{required} and \code{optional} task id values.
}
\description{
Get task id values for a given round
}
\examples{
con <- connect_hub(system.file("hub_1", package = "hubUtils"))
get_task_id_vals(con, task_id = "location")
con <- connect_hub(system.file("scnr_hub_1", package = "hubUtils"))
get_task_id_vals(con, round_id = "round-1", task_id = "location")
get_task_id_vals(con, round_id = "round-2", task_id = "age_group")
get_task_id_vals(con,
  round_id = "round-1", task_id = "location",
  flatten = FALSE
)
}
\seealso{
Other hub-metadata: 
\code{\link{get_round_ids}()},
\code{\link{get_task_ids}()}
}
\concept{hub-metadata}
