% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-metadata.R
\name{get_task_ids}
\alias{get_task_ids}
\title{Get task ids for a given round_id}
\usage{
get_task_ids(con, round_id = NULL)
}
\arguments{
\item{con}{a \code{hub_connection} class object.}

\item{round_id}{Character string.  A round id name. If rounds vary by a variable in hub,
argument is ignored and can be left \code{NULL} (default).}
}
\value{
A character vector of task ids.
}
\description{
Get task ids for a given round_id
}
\examples{
con <- connect_hub(system.file("hub_1", package = "hubUtils"))
get_task_ids(con)
con <- connect_hub(system.file("scnr_hub_1", package = "hubUtils"))
get_task_ids(con, round_id = "round-1")
get_task_ids(con, round_id = "round-2")
}
\seealso{
Other hub-metadata: 
\code{\link{get_round_ids}()},
\code{\link{get_task_id_vals}()}
}
\concept{hub-metadata}
