% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-arrow.R
\name{gs_bucket}
\alias{gs_bucket}
\title{Connect to a Google Cloud Storage (GCS) bucket}
\value{
A \code{SubTreeFileSystem} containing an \code{GcsFileSystem} and the bucket's
relative path. Note that this function's success does not guarantee that you
are authorized to access the bucket's contents.
}
\description{
See \code{\link[arrow:gs_bucket]{arrow::gs_bucket()}} for details.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bucket <- gs_bucket("voltrondata-labs-datasets")
\dontshow{\}) # examplesIf}
}
