% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hub_con_output}
\alias{hub_con_output}
\title{Example Hub model output data}
\format{
A \code{tbl} with 92 rows and 8 columns:
\itemize{
\item \code{forecast_date}: Origin date of the forecast.
\item \code{horizon}: Forecast horizon relative to the \code{forecast_date}.
\item \code{target}: Target variable.
\item \code{location}: Location of the forecast.
\item \code{output_type}: Output type of forecast.
\item \code{output_type_id}: Forecast output type level/identifier. In this case,
quantile level.
\item \code{value}: Forecast value.
\item \code{model_id}: Model identifier.
}
}
\usage{
hub_con_output
}
\description{
A subset of model output data accessed using \code{hubData} from the simple example
hub contained in the \code{hubUtils} package. The subset consists of \code{"quantile"} output
type data for \code{"US"} location and the most recent forecast date.
}
\keyword{datasets}
