% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hub_locations_eu}
\alias{hub_locations_eu}
\title{Available European locations}
\format{
A data frame with 32 rows and 2 columns:
\describe{
\item{location_name}{Name of the location}
\item{location}{Location abbreviation}
}
}
\usage{
hub_locations_eu
}
\description{
Data set with available European locations used in ECDC hubs
}
\keyword{datasets}
