% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hub_locations_us}
\alias{hub_locations_us}
\title{Available US locations}
\format{
A data frame with 3202 rows and 5 columns:
\describe{
\item{fips}{FIPS code}
\item{location_name}{Location name}
\item{geo_type}{Type of location}
\item{geo_value}{Location abbreviation or FIPS code}
\item{abbreviation}{Corresponding state abbrevaition}
}
}
\usage{
hub_locations_us
}
\description{
Data set with available locations for a US hub
}
\keyword{datasets}
