% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-url.R
\name{is_url}
\alias{is_url}
\title{Determine if a string is a URL}
\usage{
is_url(x)
}
\arguments{
\item{x}{character string to check if it is a URL. Must contain a protocol to
be considered a URL.}
}
\value{
Logical. \code{TRUE} if \code{x} is a URL, \code{FALSE} otherwise.
}
\description{
Determine if a string is a URL
}
\examples{
is_url("https://hubverse.io")
is_url("www.hubverse.io")
}
