% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-url.R
\name{is_valid_url}
\alias{is_valid_url}
\title{Determine if a URL is valid and reachable}
\usage{
is_valid_url(url)
}
\arguments{
\item{url}{character string of the URL to check.}
}
\value{
Logical. \code{TRUE} if the URL is valid and reachable, \code{FALSE} otherwise.
}
\description{
Determine if a URL is valid and reachable
}
\examples{
is_valid_url("https://docs.hubverse.io")
is_valid_url("https://docs.hubverse.io/invalid")
}
