% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-config-file.R
\name{launch_pretty_errors_report}
\alias{launch_pretty_errors_report}
\title{Print a concise and informative version of validation errors table.}
\usage{
launch_pretty_errors_report(x)
}
\arguments{
\item{x}{output of \link{validate_config}.}
}
\value{
prints the errors attribute of x in an informative format to the viewer. Only
available in interactive mode.
}
\description{
Print a concise and informative version of validation errors table.
}
\seealso{
validate_config

Other schema-validation: 
\code{\link{validate_config}()}
}
\concept{schema-validation}
