% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_model_id.R
\name{merge_model_id}
\alias{merge_model_id}
\alias{split_model_id}
\title{Merge/Split model output tbl \code{model_id} column}
\usage{
merge_model_id(tbl, sep = "-")

split_model_id(tbl, sep = "-")
}
\arguments{
\item{tbl}{a \code{data.frame} or \code{tibble} of model output data returned from a query
to a \verb{<hub_connection>} object.}

\item{sep}{character string. Character used as separator when concatenating
\code{team_abbr} and \code{model_abbr} values into a single \code{model_id} string or splitting
\code{model_id} into component \code{team_abbr} and \code{model_abbr}. When splitting, if
multiple instances of the separator exist in a \code{model_id} stringing,
splitting occurs occurs on the first instance.}
}
\value{
\code{tbl} with either \code{team_abbr} and \code{model_abbr} merged into a single \code{model_id}
column or \code{model_id} split into columns \code{team_abbr} and \code{model_abbr}.
}
\description{
Merge/Split model output tbl \code{model_id} column
}
\section{Functions}{
\itemize{
\item \code{merge_model_id()}: merge \code{team_abbr} and \code{model_abbr} into a single
\code{model_id} column.

\item \code{split_model_id()}: split \code{model_id} column into separate \code{team_abbr}
and \code{model_abbr} columns.

}}
\examples{
hub_con <- connect_hub(system.file("testhubs/flusight", package = "hubUtils"))
tbl <- hub_con \%>\%
    dplyr::filter(output_type == "quantile", location == "US") \%>\%
    dplyr::collect() \%>\%
    dplyr::filter(forecast_date == max(forecast_date))
tbl_split <- split_model_id(tbl)
tbl_split

# Merge model_id
tbl_merged <- merge_model_id(tbl_split)
tbl_merged

# Split / Merge using custom separator
tbl_sep <- tbl
tbl_sep$model_id <- gsub("-", "_", tbl_sep$model_id)
tbl_sep <- split_model_id(tbl_sep, sep = "_")
tbl_sep
tbl_sep <- merge_model_id(tbl_sep, sep = "_")
tbl_sep
}
