% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub-connection.R
\name{print.hub_connection}
\alias{print.hub_connection}
\title{Print a \verb{<hub_connection>} S3 class object}
\usage{
\method{print}{hub_connection}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A \verb{<hub_connection>} S3 class object.}

\item{verbose}{Logical. Whether to print the full structure of the object.
Defaults to \code{FALSE}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print a \verb{<hub_connection>} S3 class object
}
\examples{
hub_path <- system.file("testhubs/simple", package = "hubUtils")
hub_con <- connect_hub(hub_path)
hub_con
print(hub_con)
print(hub_con, verbose = TRUE)
}
