% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubmeta-read.R
\name{read_hubmeta}
\alias{read_hubmeta}
\title{Read a \code{hubmeta} file}
\usage{
read_hubmeta(path, drop_defs = TRUE)
}
\arguments{
\item{path}{path to the \verb{hubmeta file}.}

\item{drop_defs}{logical. Whether to drop the \verb{$defs} element of the metadata
(where applicable).}
}
\value{
a list of hub metadata.
}
\description{
Read a \code{hubmeta} file of hub metadata contained in \code{JSON} or \code{YAML} format.
Where applicable, the function also substitutes any \verb{$ref} element with values
contained in the element in \verb{$defs} defined by the value of the \verb{$ref} element.
}
\examples{
# Read a JSON `hubmeta` file
json_path <- system.file("hubmeta/scnr_hubmeta_ref.json", package = "hubUtils")
read_hubmeta(json_path)
# Read a JSON `hubmeta` file and keep `$defs` element
yml_path <- system.file("hubmeta/scnr_hubmeta_ref.yml", package = "hubUtils")
read_hubmeta(yml_path, drop_defs = FALSE)
}
