% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_metadata.R
\name{read_model_metadata}
\alias{read_model_metadata}
\title{Read a model metadata file into R}
\usage{
read_model_metadata(path)
}
\arguments{
\item{path}{A character string of the path to a local model metadata YAML
file.}
}
\value{
The contents of the metadata file as a named list, or \code{NULL} for an
empty file.
}
\description{
Reads a model metadata YAML file into a named list.
}
\details{
Every array in the file is returned as a list, including an array with a
single element. For example, \verb{methods: ["one"]} is returned as
\code{list("one")} and not as the string \code{"one"}.
}
\examples{
hub_path <- system.file("testhubs/simple", package = "hubUtils")
path <- fs::path(hub_path, "model-metadata", "team1-goodmodel.yaml")
read_model_metadata(path)
}
