% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-arrow.R
\name{s3_bucket}
\alias{s3_bucket}
\title{Connect to an AWS S3 bucket}
\value{
A \code{SubTreeFileSystem} containing an \code{S3FileSystem} and the bucket's
relative path. Note that this function's success does not guarantee that you
are authorized to access the bucket's contents.
}
\description{
See \code{\link[arrow:s3_bucket]{arrow::s3_bucket()}} for details.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bucket <- s3_bucket("voltrondata-labs-datasets")
\dontshow{\}) # examplesIf}
}
