% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{std_colnames}
\alias{std_colnames}
\title{Hubverse model output standard column names}
\format{
An object of class \code{character} of length 4.
}
\usage{
std_colnames
}
\description{
A named character string of standard column names used in hubverse model output data files.
The terms currently used for standard column names in the hubverse are English.
In future, however, this could be expanded to provide the basis for hub
terminology localisation.
}
\keyword{datasets}
