% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-task_ids.R
\name{subset_task_id_names}
\alias{subset_task_id_names}
\title{Subset a vector of column names to only include task IDs}
\usage{
subset_task_id_names(x)
}
\arguments{
\item{x}{character vector of column names}
}
\value{
a character vector of task ID names
}
\description{
Subset a vector of column names to only include task IDs
}
\examples{
x <- c(
  "origin_date", "horizon", "target_date",
  "location", "output_type", "output_type_id", "value"
)
subset_task_id_names(x)
}
