% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-config-file.R
\name{validate_config}
\alias{validate_config}
\title{Validate a hub config file against a Infectious Disease Modeling Hubs schema}
\usage{
validate_config(
  hub_path = ".",
  config = c("tasks", "admin"),
  config_path = NULL,
  schema_version = "from_config",
  branch = "main",
  pretty_errors = FALSE
)
}
\arguments{
\item{hub_path}{Path to a local hub directory.}

\item{config}{Name of config file to validate. One of \code{"tasks"} or \code{"admin"}.}

\item{config_path}{Defaults to \code{NULL} which assumes all config files are in
the \code{hub-config} directory in the root of hub directory. Argument \code{config_path}
can be used to override default by providing a path to the config file to be
validated.}

\item{schema_version}{Character string specifying the json schema version to
be used for validation. The default value \code{"from_config"} will use the version
specified in the \code{schema_version} property of the config file. \code{"latest"} will
use the latest version available in the Infectious Disease Modeling Hubs
\href{https://github.com/Infectious-Disease-Modeling-Hubs/schemas}{schemas repository}.
Alternartively, a specific version of a schema (e.g. \code{"v0.0.1"})  can be specified.}

\item{branch}{The branch of the Infectious Disease Modeling Hubs
\href{https://github.com/Infectious-Disease-Modeling-Hubs/schemas}{schemas repository}
from which to fetch schema. Defaults to \code{"main"}.}

\item{pretty_errors}{Whether to launch an informative pretty table of any
validation errors in the Viewer panel. Will only launch run in an interactive
environment.}
}
\value{
Returns the result of validation. If validation is successful, will
return \code{TRUE}. If any validation errors are detected, returns \code{FALSE} with
details of errors appended as a data.frame to an \code{errors} attribute. To access
the errors table use \code{attr(x, "errors")} where \code{x} is the output of the function.
}
\description{
Validate a hub config file against a Infectious Disease Modeling Hubs schema
}
\examples{
# Valid config file
validate_config(
  hub_path = system.file(
    "testhubs/simple/",
    package = "hubUtils"
  ),
  config = "tasks"
)
# Config file with errors
config_path <- system.file("error-schema/tasks-errors.json",
  package = "hubUtils"
)
validate_config(config_path = config_path, config = "tasks")
}
