% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_metadata_schema.R
\name{validate_model_metadata_schema}
\alias{validate_model_metadata_schema}
\title{Validate model-metadata-schema config file}
\usage{
validate_model_metadata_schema(hub_path = ".")
}
\arguments{
\item{hub_path}{Path to a local hub directory.}
}
\value{
Returns the result of validation. If validation is successful, will
return \code{TRUE}. If any validation errors are detected, returns \code{FALSE} with
details of errors appended as a data.frame to an \code{errors} attribute.
To access
the errors table use \code{attr(x, "errors")} where \code{x} is the output of the function.

You can print a more concise and easier to view version of an errors table with
\code{\link[=view_config_val_errors]{view_config_val_errors()}}.
}
\description{
Validate model-metadata-schema config file
}
\details{
Checks that a \code{model-metadata-schema.json} config file exists in \code{hub-config},
can be successfully parsed and contains at least either a \code{model_id} property or
both \code{team_abbr} and \code{model_abbr} properties.
}
\examples{
validate_model_metadata_schema(
  hub_path = system.file(
    "testhubs/simple/",
    package = "hubUtils"
  )
)
}
