% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_model_out_tbl.R
\name{validate_model_out_tbl}
\alias{validate_model_out_tbl}
\title{Validate a \code{model_out_tbl} object.}
\usage{
validate_model_out_tbl(tbl)
}
\arguments{
\item{tbl}{a \code{model_out_tbl} S3 class object.}
}
\value{
If valid, returns a \code{model_out_tbl} class object. Otherwise, throws an
error.
}
\description{
Validate a \code{model_out_tbl} object.
}
\examples{
md_out <- as_model_out_tbl(hub_con_output)
validate_model_out_tbl(md_out)
}
