% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_config_val_errors.R
\name{view_config_val_errors}
\alias{view_config_val_errors}
\title{Print a concise and informative version of validation errors table.}
\usage{
view_config_val_errors(x)
}
\arguments{
\item{x}{output of \link{validate_config}.}
}
\value{
prints the errors attribute of x in an informative format to the viewer. Only
available in interactive mode.
}
\description{
Print a concise and informative version of validation errors table.
}
\seealso{
validate_config

Other schema-validation: 
\code{\link{get_schema_url}()},
\code{\link{get_schema_valid_versions}()},
\code{\link{get_schema}()},
\code{\link{validate_config}()}
}
\concept{schema-validation}
