% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_to_hub_schema.R
\name{coerce_to_hub_schema}
\alias{coerce_to_hub_schema}
\alias{coerce_to_character}
\title{Coerce data.frame/tibble column data types to hub schema data types or character.}
\usage{
coerce_to_hub_schema(
  tbl,
  config_tasks,
  skip_date_coercion = FALSE,
  as_arrow_table = FALSE
)

coerce_to_character(tbl, as_arrow_table = FALSE)
}
\arguments{
\item{tbl}{a model output data.frame/tibble}

\item{config_tasks}{a list version of the content's of a hub's \code{tasks.json}
config file created using function \code{\link[=read_config]{read_config()}}.}

\item{skip_date_coercion}{Logical. Whether to skip coercing dates. This can be faster,
especially for larger \code{tbl}s.}

\item{as_arrow_table}{Logical. Whether to return an arrow table. Defaults to \code{FALSE}.}
}
\value{
\code{tbl} with column data types coerced to hub schema data types or character.
if \code{as_arrow_table = TRUE}, output is also converted to arrow table.
}
\description{
Coerce data.frame/tibble column data types to hub schema data types or character.
}
\section{Functions}{
\itemize{
\item \code{coerce_to_hub_schema()}: coerce columns to hub schema data types.

\item \code{coerce_to_character()}: coerce all columns to character

}}
