% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub-connection.R
\name{connect_hub}
\alias{connect_hub}
\title{Connect to a Modeling Hub}
\usage{
connect_hub(
  hub_path,
  model_output_dir = NULL,
  file_format = c("csv", "parquet", "arrow")
)
}
\arguments{
\item{hub_path}{Path to a Modeling Hub directory. The hub must be fully
configured with valid \code{admin.json} and \code{tasks.json} files within the \code{hub-config}
directory.}

\item{model_output_dir}{Path to the directory containing model output data.
This argument can be used to access data directly from an appropriately set up
model output directory which is not part of a fully configured hub.}

\item{file_format}{The file format model output files are stored in.
For connection to a fully configured hub, accessed through \code{hub_path},
\code{file_format} is inferred from the hub's \code{file_format} configuration in
\code{admin.json} and is ignored by default.
If supplied, it will override hub configuration setting.}
}
\value{
an S3 object of class \verb{<hub-connection>} connected to the data in the
model-output directory via an Apache arrow \code{FileSystemDataset} connection. The
connection can be used to extract data using \code{dplyr} custom queries. The object
also contains metadata about the modeling hub.
}
\description{
Connect to a Modeling Hub of model output directory.
}
\examples{
# Connect to a simple forecasting Hub.
hub_path <- system.file("testhubs/simple", package = "hubUtils")
hub_con <- connect_hub(hub_path)
hub_con
model_output_dir <- system.file("testhubs/simple/model-output", package = "hubUtils")
model_output_con <- connect_hub(model_output_dir = model_output_dir)
model_output_con
# Access data
library(dplyr)
hub_con \%>\%
  filter(
    origin_date == "2022-10-08",
    horizon == 2
  ) \%>\%
  collect()
model_output_con \%>\%
  filter(
    origin_date == "2022-10-08",
    horizon == 2
  ) \%>\%
  collect()
}
