% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hub_schema.R
\name{create_hub_schema}
\alias{create_hub_schema}
\title{Create a Hub arrow schema}
\usage{
create_hub_schema(
  config_tasks,
  format = c("csv", "parquet", "arrow"),
  partitions = list(model = arrow::utf8())
)
}
\arguments{
\item{config_tasks}{a list version of the content's of a hub's \code{tasks.json}
config file created using function \code{\link[=read_config]{read_config()}}.}

\item{format}{the hub file format.}

\item{partitions}{a named list specifying the arrow data types of any
partitioning column. Only relevant for opening \code{"parquet"} or \code{"arrow"} format datasets.
Ignored for all other formats.}
}
\value{
an arrow schema object that can be used to define column datatypes when
opening model output data. If the schema is being created for a \code{parquet}/\code{arrow} hub,
the partitioning variable data type is included in the schema while for \code{csv} hubs
the partitioning variable is not as it causes errors when opening.
}
\description{
Create an arrow schema from a the \code{tasks.json} config file for using when
opening an arrow dataset.
}
\examples{
origin_path <- system.file("testhubs/simple", package = "hubUtils")
config_tasks <- hubUtils:::read_config(origin_path, "tasks")
schema_csv <- create_hub_schema(config_tasks, format = "csv")
con_csv <- arrow::open_dataset(
    origin_path, format = "csv",
    partitioning = "model",
    col_types = schema_csv,
    factory_options = list(exclude_invalid_files = TRUE))
con_csv

schema_parquet <- create_hub_schema(config_tasks, format = "parquet")
con_parquet <- arrow::open_dataset(
    origin_path, format = "parquet",
    partitioning = "model",
    schema = schema_parquet,
    factory_options = list(exclude_invalid_files = TRUE))
con_parquet

hub_con <- arrow::open_dataset(
    sources = list(
        con_csv,
        con_parquet)
)
hub_con
}
