% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hub_schema.R
\name{create_hub_schema}
\alias{create_hub_schema}
\title{Create a Hub arrow schema}
\usage{
create_hub_schema(
  config_tasks,
  partitions = list(model_id = arrow::utf8()),
  output_type_id_datatype = c("auto", "character", "double", "integer", "logical",
    "Date"),
  r_schema = FALSE
)
}
\arguments{
\item{config_tasks}{a list version of the content's of a hub's \code{tasks.json}
config file created using function \code{\link[=read_config]{read_config()}}.}

\item{partitions}{a named list specifying the arrow data types of any
partitioning column.}

\item{output_type_id_datatype}{character string. One of \code{"auto"}, \code{"character"},
\code{"double"}, \code{"integer"}, \code{"logical"}, \code{"Date"}. Defaults to \code{"auto"} indicating
that \code{output_type_id} will be determined automatically from the \code{tasks.json}
config file. Other data type values can be used to override automatic determination.
Note that attempting to coerce \code{output_type_id} to a data type that is not possible
(e.g. trying to coerce to \code{"double"} when the data contains \code{"character"} values)
will likely result in an error or potentially unexpected behaviour so use with
care.}

\item{r_schema}{Logical. If \code{FALSE} (default), return an \code{\link[arrow:schema]{arrow::schema()}} object.
If \code{TRUE}, return a character vector of R data types.}
}
\value{
an arrow schema object that can be used to define column datatypes when
opening model output data. If \code{r_schema = TRUE}, a character vector of R data types.
}
\description{
Create an arrow schema from a \code{tasks.json} config file. For use when
opening an arrow dataset.
}
\examples{
hub_path <- system.file("testhubs/simple", package = "hubUtils")
config_tasks <- read_config(hub_path, "tasks")
schema <- create_hub_schema(config_tasks)
}
