% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_model_out_val_grid.R
\name{expand_model_out_val_grid}
\alias{expand_model_out_val_grid}
\title{Create expanded grid of valid task ID and output type value combinations}
\usage{
expand_model_out_val_grid(
  config_tasks,
  round_id,
  required_vals_only = FALSE,
  all_character = FALSE,
  as_arrow_table = FALSE,
  bind_model_tasks = TRUE
)
}
\arguments{
\item{config_tasks}{a list version of the content's of a hub's \code{tasks.json}
config file, accessed through the \code{"config_tasks"} attribute of a \verb{<hub_connection>}
object or function \code{\link[=read_config]{read_config()}}.}

\item{round_id}{Character string. Round identifier. If the round is set to
\code{round_id_from_variable: true}, IDs are values of the task ID defined in the round's
\code{round_id} property of \code{config_tasks}.
Otherwise should match round's \code{round_id} value in config. Ignored if hub
contains only a single round.}

\item{required_vals_only}{Logical. Whether to return only combinations of
Task ID and related output type ID required values.}

\item{all_character}{Logical. Whether to return all character column.}

\item{as_arrow_table}{Logical. Whether to return an arrow table. Defaults to \code{FALSE}.}

\item{bind_model_tasks}{Logical. Whether to bind expanded grids of
values from multiple modeling tasks into a single tibble/arrow table or
return a list.}
}
\value{
If \code{bind_model_tasks = TRUE} (default) a tibble or arrow table
containing all possible task ID and related output type ID
value combinations. If \code{bind_model_tasks = FALSE}, a list containing a
tibble or arrow table for each round modeling task.

Columns are coerced to data types according to the hub schema,
unless \code{all_character = TRUE}. If \code{all_character = TRUE}, all columns are returned as
character which can be faster when large expanded grids are expected.
If \code{required_vals_only = TRUE}, values are limited to the combinations of required
values only.
}
\description{
Create expanded grid of valid task ID and output type value combinations
}
\details{
When a round is set to \code{round_id_from_variable: true},
the value of the task ID from which round IDs are derived (i.e. the task ID
specified in \code{round_id} property of \code{config_tasks}) is set to the value of the
\code{round_id} argument in the returned output.
}
\examples{
hub_con <- connect_hub(
  system.file("testhubs/flusight", package = "hubUtils")
)
config_tasks <- attr(hub_con, "config_tasks")
expand_model_out_val_grid(config_tasks, round_id = "2023-01-02")
expand_model_out_val_grid(
  config_tasks,
  round_id = "2023-01-02",
  required_vals_only = TRUE
)
# Specifying a round in a hub with multiple round configurations.
hub_con <- connect_hub(
  system.file("testhubs/simple", package = "hubUtils")
)
config_tasks <- attr(hub_con, "config_tasks")
expand_model_out_val_grid(config_tasks, round_id = "2022-10-01")
# Later round_id maps to round config that includes additional task ID 'age_group'.
expand_model_out_val_grid(config_tasks, round_id = "2022-10-29")
# Coerce all columns to character
expand_model_out_val_grid(config_tasks,
  round_id = "2022-10-29",
  all_character = TRUE
)
# Return arrow table
expand_model_out_val_grid(config_tasks,
  round_id = "2022-10-29",
  all_character = TRUE,
  as_arrow_table = TRUE
)
}
