% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_model_out_val_grid.R
\name{expand_model_out_val_grid}
\alias{expand_model_out_val_grid}
\title{Create expanded grid of valid task ID and output type value combinations}
\usage{
expand_model_out_val_grid(config_tasks, round_id = NULL, required_only = FALSE)
}
\arguments{
\item{config_tasks}{a list version of the content's of a hub's \code{tasks.json}
config file, accessed through the \code{"config_tasks"} attribute of a \verb{<hub_connection>}
object or function \code{\link[=read_config]{read_config()}}.}

\item{round_id}{Character string. Round identifier. If the round is set to
\code{round_id_from_variable: true}, IDs are values of the task ID defined in
\code{round_id}. Otherwise should match config \code{round_id} values. Ignored if hub
contains only a single round.}

\item{required_only}{Logical. Whether to return combinations of required values only.}
}
\value{
a tibble containing all possible task ID and related output type ID
value combinations.
}
\description{
Create expanded grid of valid task ID and output type value combinations
}
\examples{
hub_con <- connect_hub(system.file("testhubs/flusight", package = "hubUtils"))
config_tasks <- attr(hub_con, "config_tasks")
expand_model_out_val_grid(config_tasks)
expand_model_out_val_grid(config_tasks, required_only = TRUE)
# Specifying a round in a hub with multiple rounds
hub_con <- connect_hub(system.file("testhubs/simple", package = "hubUtils"))
config_tasks <- attr(hub_con, "config_tasks")
expand_model_out_val_grid(config_tasks, round_id = "2022-10-01")
expand_model_out_val_grid(config_tasks, round_id = "2022-10-29")
}
