% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-schema.R
\name{extract_schema_version}
\alias{extract_schema_version}
\title{Extract the schema version from a schema \code{id} or config \code{schema_version} property
character string}
\usage{
extract_schema_version(id)
}
\arguments{
\item{id}{A schema \code{id} or config \code{schema_version} property character string.}
}
\value{
The schema version number as a character string.
}
\description{
Extract the schema version from a schema \code{id} or config \code{schema_version} property
character string
}
