% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-config.R
\name{get_derived_task_ids_config}
\alias{get_derived_task_ids_config}
\alias{get_config_derived_task_ids}
\title{Get hub or round level derived task IDs}
\usage{
get_derived_task_ids_config(
  config_tasks,
  round_id = NULL,
  call = rlang::current_env()
)

get_config_derived_task_ids(
  config_tasks,
  round_id = NULL,
  call = rlang::current_env()
)
}
\arguments{
\item{config_tasks}{a list version of the content's of a hub's \code{tasks.json}
config file, accessed through the \code{"config_tasks"} attribute of a \verb{<hub_connection>}
object or function \code{\link[=read_config]{read_config()}}.}

\item{round_id}{Character string. Round identifier. If the round is set to
\code{round_id_from_variable: true}, IDs are values of the task ID defined in the
round's \code{round_id} property of \code{config_tasks}.
Otherwise should match round's \code{round_id} value in config.}

\item{call}{The execution environment of the function to name in error and
warning messages. By default, messages name this function. Supply another
environment, such as a wrapper function's environment, to name that
function instead.}
}
\value{
a character vector of hub or round level derived task ID names.
If \code{round_id} is \code{NULL} or the round does not have a round level
\code{derived_task_ids} setting, returns the hub level \code{derived_task_ids}
setting.
}
\description{
\code{get_config_derived_task_ids()} is an alias of
\code{get_derived_task_ids_config()}.
}
\examples{
hub_path <- system.file("testhubs/v4/flusight", package = "hubUtils")
config_tasks <- read_config(hub_path, "tasks")
get_derived_task_ids_config(config_tasks)
get_derived_task_ids_config(config_tasks, round_id = "2023-05-08")
}
\seealso{
\code{\link[=get_hub_derived_task_ids]{get_hub_derived_task_ids()}} to get derived task IDs from a hub
path.
}
