% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get_hub.R
\name{get_hub_timezone}
\alias{get_hub_timezone}
\alias{get_hub_model_output_dir}
\alias{get_hub_file_formats}
\title{Get hub configuration fields}
\usage{
get_hub_timezone(hub_path)

get_hub_model_output_dir(hub_path)

get_hub_file_formats(hub_path, round_id = NULL)
}
\arguments{
\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[arrow:s3_bucket]{arrow::s3_bucket()}}
or \code{\link[arrow:gs_bucket]{arrow::gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.}

\item{round_id}{Character string. Round identifier. If the round is set to
\code{round_id_from_variable: true}, IDs are values of the task ID defined in the
round's \code{round_id} property of \code{config_tasks}.
Otherwise should match round's \code{round_id} value in config. Ignored if hub
contains only a single round.}
}
\value{
\itemize{
\item \code{get_hub_timezone}: The timezone of the hub
}

\itemize{
\item \code{get_hub_model_output_dir}: The model output directory name
}

\itemize{
\item \code{get_hub_file_formats}: The accepted hub or round file formats
}
}
\description{
Get hub configuration fields
}
\section{Functions}{
\itemize{
\item \code{get_hub_timezone()}: Get the hub timezone

\item \code{get_hub_model_output_dir()}: Get the model output directory name

\item \code{get_hub_file_formats()}: Get the hub or round file formats

}}
\examples{
hub_path <- system.file("testhubs", "flusight", package = "hubUtils")
get_hub_timezone(hub_path)
get_hub_model_output_dir(hub_path)
get_hub_file_formats(hub_path)
get_hub_file_formats(hub_path, "2022-12-12")
}
