% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_id_utils.R
\name{get_round_idx}
\alias{get_round_idx}
\alias{get_round_ids}
\title{Utilities for accessing round ID metadata}
\usage{
get_round_idx(config_tasks, round_id = NULL)

get_round_ids(config_tasks, flatten = TRUE)
}
\arguments{
\item{config_tasks}{a list version of the content's of a hub's \code{tasks.json}
config file, accessed through the \code{"config_tasks"} attribute of a \verb{<hub_connection>}
object or function \code{\link[=read_config]{read_config()}}.}

\item{round_id}{Character string. Round identifier. If the round is set to
\code{round_id_from_variable: true}, IDs are values of the task ID defined in the round's
\code{round_id} property of \code{config_tasks}.
Otherwise should match round's \code{round_id} value in config. Ignored if hub
contains only a single round. Must be specified if \code{config_tasks} contains
more than one round.}

\item{flatten}{Logical. Whether to flatten output to character vector of
round IDs.}
}
\description{
Utilities for accessing round ID metadata
}
\section{Functions}{
\itemize{
\item \code{get_round_idx()}: Get an integer index of the element in
\code{config_tasks$rounds} that a character round identifier maps to.

\item \code{get_round_ids()}: Get a list or character vector of hub round IDs.

}}
\examples{
hub_con <- connect_hub(system.file("testhubs/simple", package = "hubUtils"))
config_tasks <- attr(hub_con, "config_tasks")
# Get round IDs
get_round_ids(config_tasks)
get_round_ids(config_tasks, flatten = FALSE)
get_round_idx(config_tasks, "2022-10-01")
get_round_idx(config_tasks, "2022-10-29")
}
