% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config.R
\name{get_schema_url}
\alias{get_schema_url}
\title{Get the json schema download URL for a given config file version}
\usage{
get_schema_url(config = c("tasks", "admin", "model"), version, branch = "main")
}
\arguments{
\item{config}{Name of config file to validate. One of \code{"tasks"} or \code{"admin"}.}

\item{version}{A valid version of Infectious Disease Modeling Hubs
\href{https://github.com/Infectious-Disease-Modeling-Hubs/schemas}{schema}
(e.g. \code{"v0.0.1"}).}

\item{branch}{The branch of the Infectious Disease Modeling Hubs
\href{https://github.com/Infectious-Disease-Modeling-Hubs/schemas}{schemas repository}
from which to fetch schema. Defaults to \code{"main"}.}
}
\value{
The json schema download URL for a given config file version.
}
\description{
Get the json schema download URL for a given config file version
}
\examples{
get_schema_url(config = "tasks", version = "v0.0.0.9")
}
\seealso{
Other functions supporting config file validation: 
\code{\link{get_schema}()},
\code{\link{get_schema_valid_versions}()},
\code{\link{validate_config}()},
\code{\link{validate_hub_config}()},
\code{\link{view_config_val_errors}()}
}
\concept{functions supporting config file validation}
