% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-schema.R
\name{get_schema_valid_versions}
\alias{get_schema_valid_versions}
\title{Get a vector of valid schema version}
\usage{
get_schema_valid_versions(branch = "main")
}
\arguments{
\item{branch}{The branch of the Infectious Disease Modeling Hubs
\href{https://github.com/hubverse-org/schemas}{schemas repository}
from which to fetch schema. Defaults to \code{"main"}.}
}
\value{
a character vector of valid versions of Infectious Disease Modeling Hubs
\href{https://github.com/hubverse-org/schemas}{schema}.
}
\description{
Get a vector of valid schema version
}
\examples{
get_schema_valid_versions()
}
\seealso{
Other functions supporting config file validation: 
\code{\link{get_schema}()},
\code{\link{get_schema_url}()}
}
\concept{functions supporting config file validation}
