% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-schema-versions.R
\name{get_version_config}
\alias{get_version_config}
\alias{get_version_file}
\alias{get_version_hub}
\title{Get hub config schema versions}
\usage{
get_version_config(config)

get_version_file(config_path)

get_version_hub(hub_path, config_type = c("tasks", "admin"))
}
\arguments{
\item{config}{List representation of the JSON config file.}

\item{config_path}{path to JSON config file}

\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[arrow:s3_bucket]{arrow::s3_bucket()}}
or \code{\link[arrow:gs_bucket]{arrow::gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.}

\item{config_type}{Character vector specifying the type of config file to read.
One of "tasks" or "admin". Default is "tasks".}
}
\value{
The schema version number as a character string.
}
\description{
Get hub config schema versions
}
\section{Functions}{
\itemize{
\item \code{get_version_config()}: Get schema version from config list representation.

\item \code{get_version_file()}: Get schema version from config file at specific path.

\item \code{get_version_hub()}: Get schema version from config file at specific path.

}}
\examples{
config <- read_config_file(
  system.file("config", "tasks.json", package = "hubUtils")
)
get_version_config(config)
config_path <- system.file("config", "tasks.json", package = "hubUtils")
get_version_file(config_path)
hub_path <- system.file("testhubs/simple", package = "hubUtils")
get_version_hub(hub_path)
get_version_hub(hub_path, "admin")
}
