% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v3-schema-utils.R
\name{is_v3_config_file}
\alias{is_v3_config_file}
\title{Is config file using v3.0.0 schema?}
\usage{
is_v3_config_file(config_path)
}
\arguments{
\item{config_path}{Path to the config file.}
}
\value{
Logical, whether the config file is using v3.0.0 schema.
}
\description{
Is config file using v3.0.0 schema?
}
\examples{
config_path <- system.file("config", "tasks.json", package = "hubUtils")
is_v3_config_file(config_path)
}
