% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_config}
\alias{read_config}
\title{Read a hub config file into R}
\usage{
read_config(hub_path, config = c("tasks", "admin", "model-metadata-schema"))
}
\arguments{
\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[hubData:s3_bucket]{hubData::s3_bucket()}}
or \code{\link[hubData:gs_bucket]{hubData::gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.}

\item{config}{Name of config file to validate. One of \code{"tasks"} or \code{"admin"}.}
}
\value{
The contents of the config as an R list.
}
\description{
Read a hub config file into R
}
\examples{
# Read config files from local hub
hub_path <- system.file("testhubs/simple", package = "hubUtils")
read_config(hub_path, "tasks")
read_config(hub_path, "admin")

\dontshow{if (asNamespace("hubUtils")$not_rcmd_check() && requireNamespace("hubData", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Read config file from AWS S3 bucket hub
hub_path <- hubData::s3_bucket("hubverse/hubutils/testhubs/simple/")
read_config(hub_path, "admin")
\dontshow{\}) # examplesIf}
}
