% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_config_file}
\alias{read_config_file}
\title{Read a JSON config file from a path}
\usage{
read_config_file(config_path)
}
\arguments{
\item{config_path}{Character string. Path to JSON config file.}
}
\value{
The contents of the config file as an R list. If possible, the output is
further converted to a \verb{<config>} class object before returning.
}
\description{
Read a JSON config file from a path
}
\examples{
read_config_file(system.file("config", "tasks.json", package = "hubUtils"))
}
