% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_config_file}
\alias{read_config_file}
\title{Read a JSON config file from a path}
\usage{
read_config_file(config_path, silent = TRUE)
}
\arguments{
\item{config_path}{Character string. Path to JSON config file.}

\item{silent}{Logical. If \code{TRUE}, suppress warnings. Default is \code{FALSE}.}
}
\value{
The contents of the config file as an R list. If possible, the output is
further converted to a \verb{<config>} class object before returning. Note that
\code{"model-metadata-schema"} files are never converted to a \verb{<config>} object.
}
\description{
Read a JSON config file from a path
}
\examples{
read_config_file(system.file("config", "tasks.json", package = "hubUtils"))
}
