% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_config_file}
\alias{read_config_file}
\title{Read a JSON config file from a path}
\usage{
read_config_file(config_path)
}
\arguments{
\item{config_path}{path to JSON config file}
}
\value{
a list representation of the JSON config file
}
\description{
Read a JSON config file from a path
}
\examples{
read_config_file(system.file("config", "tasks.json", package = "hubUtils"))
}
