% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_model_out_tbl.R
\name{validate_model_out_tbl}
\alias{validate_model_out_tbl}
\title{Validate a \code{model_out_tbl} object.}
\usage{
validate_model_out_tbl(tbl)
}
\arguments{
\item{tbl}{a \code{model_out_tbl} S3 class object.}
}
\value{
If valid, returns a \code{model_out_tbl} class object. Otherwise, throws an
error.
}
\description{
Validate a \code{model_out_tbl} object.
}
\examples{
if (requireNamespace("hubData", quietly = TRUE)) {
  library(dplyr)
  hub_path <- system.file("testhubs/flusight", package = "hubUtils")
  hub_con <- hubData::connect_hub(hub_path)
  md_out <- hub_con \%>\%
    filter(output_type == "quantile", location == "US") \%>\%
    collect() \%>\%
    filter(forecast_date == max(forecast_date)) \%>\%
    as_model_out_tbl()

  validate_model_out_tbl(md_out)
}
}
