% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_config_val_errors.R
\name{view_config_val_errors}
\alias{view_config_val_errors}
\title{Print a concise and informative version of validation errors table.}
\usage{
view_config_val_errors(x)
}
\arguments{
\item{x}{output of \code{\link[=validate_config]{validate_config()}}.}
}
\value{
prints the errors attribute of x in an informative format to the viewer. Only
available in interactive mode.
}
\description{
Print a concise and informative version of validation errors table.
}
\examples{
\dontrun{
config_path <- system.file("error-schema/tasks-errors.json",
  package = "hubUtils"
)
validate_config(config_path = config_path, config = "tasks") |>
  view_config_val_errors()
}
}
\seealso{
\code{\link[=validate_config]{validate_config()}}

Other functions supporting config file validation: 
\code{\link{get_schema}()},
\code{\link{get_schema_url}()},
\code{\link{get_schema_valid_versions}()},
\code{\link{validate_config}()},
\code{\link{validate_hub_config}()}
}
\concept{functions supporting config file validation}
